Move firstDayOfWeek getter to Scheduler
