+++ packages/dx-react-scheduler/docs/reference/week-view.md
@@ -29,7 +29,6 @@ Name | Type | Default | Description
 name? | string | `Week` | The view's unique identifier. Required if you use several `WeekView` plugins.
 displayName? | string |  | The view's name used in UI plugins. The default value is equal to `name`.
 excludedDays? | Array&lt;number&gt; | [] | Specifies the days of week that should not be displayed in the view. Accepts an array of zero-bazed day indexes (0 - Sunday).
-firstDayOfWeek? | number | 0 | Specifies the first day of week.
 intervalCount? | number | 1 | Multiplies the default view interval.
 cellDuration? | number | 30 | Specifies the cell's duration in minutes.
 startDayHour? | number | 0 | Specifies the start hour of the view time scale.
