+++ packages/dx-react-scheduler/src/plugins/month-view.test.tsx
@@ -122,9 +122,11 @@ describe('Month View', () => {
       monthCellsData.mockImplementation(() => expectedMonthCellsData);
       const tree = mount((
         <PluginHost>
-          {pluginDepsToComponents(defaultDeps)}
+          {pluginDepsToComponents({
+            ...defaultDeps,
+            getter: { ...defaultDeps.getter, firstDayOfWeek },
+          })}
           <MonthView
-            firstDayOfWeek={firstDayOfWeek}
             intervalCount={intervalCount}
             {...defaultProps}
           />
@@ -141,22 +143,6 @@ describe('Month View', () => {
         .toBe(intervalCount);
       expect(getComputedState(tree).viewCellsData)
         .toEqual(expectedMonthCellsData);
-    });
-
-    it('should provide the "firstDayOfWeek" getter', () => {
-      const firstDayOfWeek = 2;
-      const tree = mount((
-        <PluginHost>
-          {pluginDepsToComponents(defaultDeps)}
-          <MonthView
-            firstDayOfWeek={firstDayOfWeek}
-            {...defaultProps}
-          />
-        </PluginHost>
-      ));
-
-      expect(getComputedState(tree).firstDayOfWeek)
-        .toBe(firstDayOfWeek);
     });
 
     it('should provide the "startViewDate" getter', () => {
@@ -183,22 +169,6 @@ describe('Month View', () => {
       ));
       expect(getComputedState(tree).endViewDate)
         .toEqual(new Date('2018-08-06'));
-    });
-
-    it('should provide the "firstDayOfWeek" getter', () => {
-      const firstDayOfWeek = 2;
-      const tree = mount((
-        <PluginHost>
-          {pluginDepsToComponents(defaultDeps)}
-          <MonthView
-            firstDayOfWeek={firstDayOfWeek}
-            {...defaultProps}
-          />
-        </PluginHost>
-      ));
-
-      expect(getComputedState(tree).firstDayOfWeek)
-        .toBe(firstDayOfWeek);
     });
 
     it('should provide the "intervalCount" getter', () => {
