+++ packages/dx-react-scheduler/src/types/scheduler/scheduler.types.ts
@@ -23,4 +23,6 @@ export interface SchedulerProps {
    * the height equals that of the container component.
    * **/
   height: number | 'auto';
+  /** A number between 0 (Sunday) and 6 (Saturday) that specifies the first day of the week. */
+  firstDayOfWeek?: number;
 }
